Minor test refactoring
